landslide_app/requirements.txt
streamlit
torch
torchvision
albumentations
opencv-python
numpy
pillow
